+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This page shows a detailed overview of the changes between versions without the 
 	## **WORK IN PROGRESS**
 -->
 
+## **WORK IN PROGRESS**
+
+- Speed up attribute data conversion for migration and websocket
+
 ## 1.1.6 (2026-06-30)
 
 - Ensure that rebuilding of node structures after updates does not block the event loop for big bridges